requirements.txt
Flask
Jinja2
PyYAML
requests
scopus
scholarly
sqlitedict
